pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fr-studio"
version = "0.1.0"
description = "Frgeek古着オンラインストア向け画像加工・商品説明生成ツール"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Frgeek" },
]
dependencies = [
    "pillow==12.1.0",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "transformers==4.57.5",
    "einops==0.8.1",
    "kornia==0.8.2",
    "timm==1.0.24",
    "scipy==1.17.0",
    "scikit-learn==1.8.0",
    "peft==0.18.1",
    "datasets==4.5.0",
    "accelerate==1.12.0",
    "PySide6>=6.6.0",
    "peewee>=3.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "mypy==1.19.1",
    "ruff==0.14.13",
]

[tool.hatch.build.targets.wheel]
packages = ["src/fr_studio"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
